+++ java/src/main/java/ai/rapids/cudf/Cuda.java
@@ -259,28 +259,52 @@ public class Cuda {
   static native void freePinned(long ptr) throws CudaException;
 
   /**
-   * Copies count bytes from the memory area pointed to by src to the memory area pointed to by
-   * dst.
-   * Calling cudaMemcpy() with dst and src pointers that do not
-   * match the direction of the copy results in an undefined behavior.
+   * Copies bytes between buffers using the default CUDA stream.
+   * The copy has completed when this returns, but the memory copy could overlap with
+   * operations occurring on other streams.
+   * Specifying pointers that do not match the copy direction results in undefined behavior.
    * @param dst   - Destination memory address
    * @param src   - Source memory address
    * @param count - Size in bytes to copy
    * @param kind  - Type of transfer. {@link CudaMemcpyKind}
    */
   static void memcpy(long dst, long src, long count, CudaMemcpyKind kind) {
-    memcpy(dst, src, count, kind.getValue());
+    memcpy(dst, src, count, kind, DEFAULT_STREAM);
   }
 
-  private static native void memcpy(long dst, long src, long count, int kind) throws CudaException;
+  /**
+   * Copies bytes between buffers using the default CUDA stream.
+   * The copy has not necessarily completed when this returns, but the memory copy could
+   * overlap with operations occurring on other streams.
+   * Specifying pointers that do not match the copy direction results in undefined behavior.
+   * @param dst   - Destination memory address
+   * @param src   - Source memory address
+   * @param count - Size in bytes to copy
+   * @param kind  - Type of transfer. {@link CudaMemcpyKind}
+   */
+  static void asyncMemcpy(long dst, long src, long count, CudaMemcpyKind kind) {
+    asyncMemcpy(dst, src, count, kind, DEFAULT_STREAM);
+  }
 
   /**
    * Sets count bytes starting at the memory area pointed to by dst, with value.
+   * The operation has completed when this returns, but it could overlap with operations occurring
+   * on other streams.
    * @param dst   - Destination memory address
    * @param value - Byte value to set dst with
    * @param count - Size in bytes to set
    */
   public static native void memset(long dst, byte value, long count) throws CudaException;
+
+  /**
+   * Sets count bytes starting at the memory area pointed to by dst, with value.
+   * The operation has not necessarily completed when this returns, but it could overlap with
+   * operations occurring on other streams.
+   * @param dst   - Destination memory address
+   * @param value - Byte value to set dst with
+   * @param count - Size in bytes to set
+   */
+  public static native void asyncMemset(long dst, byte value, long count) throws CudaException;
 
   /**
    * Get the id of the current device.