+++ python/cudf/cudf/core/column/column.py
@@ -28,6 +28,7 @@ from cudf.core.dtypes import CategoricalDtype
 from cudf.utils import ioutils, utils
 from cudf.utils.dtypes import (
     check_cast_unsupported_dtype,
+    get_time_unit,
     is_categorical_dtype,
     is_list_dtype,
     is_numerical_dtype,
@@ -36,7 +37,6 @@ from cudf.utils.dtypes import (
     min_unsigned_type,
     np_to_pa_dtype,
 )
-
 from cudf.utils.utils import mask_dtype
 
 
@@ -70,12 +70,10 @@ class ColumnBase(Column, Serializable):
         )
 
     def as_frame(self):
-        from cudf.core.frame import Frame
-
         """
         Converts a Column to Frame
         """
-        return Frame({None: self.copy(deep=False)})
+        return cudf.core.frame.Frame({None: self.copy(deep=False)})
 
     @property
     def data_array_view(self):
@@ -478,7 +476,6 @@ class ColumnBase(Column, Serializable):
         return libcudf.copying.get_element(self, index).value
 
     def __getitem__(self, arg):
-        from cudf.core.column import column
 
         if isinstance(arg, Number):
             arg = int(arg)
@@ -521,9 +518,9 @@ class ColumnBase(Column, Serializable):
                 )
                 return self.take(gather_map)
         else:
-            arg = column.as_column(arg)
+            arg = as_column(arg)
             if len(arg) == 0:
-                arg = column.as_column([], dtype="int32")
+                arg = as_column([], dtype="int32")
             if pd.api.types.is_integer_dtype(arg.dtype):
                 return self.take(arg)
             if pd.api.types.is_bool_dtype(arg.dtype):
@@ -537,7 +534,6 @@ class ColumnBase(Column, Serializable):
         If value and self are of different types,
         value is coerced to self.dtype
         """
-        from cudf.core import column
 
         if isinstance(key, slice):
             key_start, key_stop, key_stride = key.indices(len(self))
@@ -562,15 +558,15 @@ class ColumnBase(Column, Serializable):
             else:
                 nelem = abs(key_stop - key_start)
         else:
-            key = column.as_column(key)
+            key = as_column(key)
             if pd.api.types.is_bool_dtype(key.dtype):
                 if not len(key) == len(self):
                     raise ValueError(
                         "Boolean mask must be of same length as column"
                     )
-                key = column.as_column(cupy.arange(len(self)))[key]
+                key = as_column(cupy.arange(len(self)))[key]
                 if hasattr(value, "__len__") and len(value) == len(self):
-                    value = column.as_column(value)[key]
+                    value = as_column(value)[key]
             nelem = len(key)
 
         if is_scalar(value):
@@ -586,7 +582,7 @@ class ColumnBase(Column, Serializable):
                     f"{nelem}"
                 )
                 raise ValueError(msg)
-            value = column.as_column(value).astype(self.dtype)
+            value = as_column(value).astype(self.dtype)
             if is_categorical_dtype(value.dtype):
                 value = value.cat().set_categories(self.categories)
                 assert self.dtype == value.dtype
@@ -695,8 +691,6 @@ class ColumnBase(Column, Serializable):
         return indices[-1]
 
     def append(self, other):
-        from cudf.core.column import as_column
-
         return ColumnBase._concat([self, as_column(other)])
 
     def quantile(self, q, interpolation, exact):
@@ -815,9 +809,8 @@ class ColumnBase(Column, Serializable):
     @ioutils.doc_to_dlpack()
     def to_dlpack(self):
         """{docstring}"""
-        from cudf.io import dlpack as dlpack
 
-        return dlpack.to_dlpack(self)
+        return cudf.io.dlpack.to_dlpack(self)
 
     @property
     def is_unique(self):
@@ -888,12 +881,15 @@ class ColumnBase(Column, Serializable):
     def astype(self, dtype, **kwargs):
         if is_categorical_dtype(dtype):
             return self.as_categorical_column(dtype, **kwargs)
+<<<<<<< HEAD
         elif pd.api.types.pandas_dtype(dtype).type in (
             np.str_,
             np.object_,
             str,
         ):
             return self.as_string_column(dtype, **kwargs)
+=======
+>>>>>>> fd94d283b3c828fa49105acc13a802072e27069c
         elif is_list_dtype(dtype):
             if not self.dtype == dtype:
                 raise NotImplementedError(
@@ -902,6 +898,13 @@ class ColumnBase(Column, Serializable):
             return self
         elif np.issubdtype(dtype, np.datetime64):
             return self.as_datetime_column(dtype, **kwargs)
+<<<<<<< HEAD
+=======
+        elif np.issubdtype(dtype, np.timedelta64):
+            return self.as_timedelta_column(dtype, **kwargs)
+        elif pd.api.types.pandas_dtype(dtype).type in (np.str_, np.object_):
+            return self.as_string_column(dtype, **kwargs)
+>>>>>>> fd94d283b3c828fa49105acc13a802072e27069c
         else:
             return self.as_numerical_column(dtype, **kwargs)
 
@@ -954,6 +957,9 @@ class ColumnBase(Column, Serializable):
     def as_datetime_column(self, dtype, **kwargs):
         raise NotImplementedError
 
+    def as_timedelta_column(self, dtype, **kwargs):
+        raise NotImplementedError
+
     def as_string_column(self, dtype, **kwargs):
         raise NotImplementedError
 
@@ -964,7 +970,7 @@ class ColumnBase(Column, Serializable):
         )
         return result
 
-    def argsort(self, ascending, na_position="last"):
+    def argsort(self, ascending=True, na_position="last"):
 
         sorted_indices = self.as_frame()._get_sorted_inds(
             ascending=ascending, na_position=na_position
@@ -1174,6 +1180,15 @@ def build_column(
             offset=offset,
             null_count=null_count,
         )
+    elif dtype.type is np.timedelta64:
+        return cudf.core.column.TimeDeltaColumn(
+            data=data,
+            dtype=dtype,
+            mask=mask,
+            size=size,
+            offset=offset,
+            null_count=null_count,
+        )
     elif dtype.type in (np.object_, np.str_):
         return cudf.core.column.StringColumn(
             mask=mask,
@@ -1349,6 +1364,7 @@ def as_column(arbitrary, nan_as_null=None, dtype=None, length=None):
                 col = utils.time_col_replace_nulls(col)
         return col
 
+<<<<<<< HEAD
     elif isinstance(arbitrary, (pa.Array, pa.ChunkedArray)):
         col = ColumnBase.from_arrow(arbitrary)
         if isinstance(col, cudf.core.column.NumericalColumn) and np.issubdtype(
@@ -1363,6 +1379,12 @@ def as_column(arbitrary, nan_as_null=None, dtype=None, length=None):
             if nan_as_null is not False:
                 col = utils.time_col_replace_nulls(col)
         if isinstance(arbitrary, pa.NullArray):
+=======
+    elif isinstance(arbitrary, pa.Array):
+        if isinstance(arbitrary, pa.StringArray):
+            data = cudf.core.column.StringColumn.from_arrow(arbitrary)
+        elif isinstance(arbitrary, pa.NullArray):
+>>>>>>> fd94d283b3c828fa49105acc13a802072e27069c
             if type(dtype) == str and dtype == "empty":
                 new_dtype = pd.api.types.pandas_dtype(
                     arbitrary.type.to_pandas_dtype()
@@ -1371,7 +1393,73 @@ def as_column(arbitrary, nan_as_null=None, dtype=None, length=None):
                 new_dtype = pd.api.types.pandas_dtype(dtype)
             col = col.astype(new_dtype)
 
+<<<<<<< HEAD
         return col
+=======
+            if is_categorical_dtype(new_dtype):
+                arbitrary = arbitrary.dictionary_encode()
+            else:
+                if nan_as_null is False:
+                    # casting a null array doesn't make nans valid
+                    # so we create one with valid nans from scratch:
+                    if new_dtype == np.dtype("object"):
+                        arbitrary = utils.scalar_broadcast_to(
+                            None, (len(arbitrary),), dtype=new_dtype
+                        )
+                    else:
+                        arbitrary = utils.scalar_broadcast_to(
+                            np.nan, (len(arbitrary),), dtype=new_dtype
+                        )
+                else:
+                    arbitrary = arbitrary.cast(np_to_pa_dtype(new_dtype))
+            data = as_column(arbitrary, nan_as_null=nan_as_null)
+        elif isinstance(arbitrary, pa.DictionaryArray):
+            data = cudf.core.column.CategoricalColumn.from_arrow(arbitrary)
+        elif isinstance(arbitrary, pa.TimestampArray):
+            data = cudf.core.column.DatetimeColumn.from_arrow(arbitrary)
+        elif isinstance(arbitrary, pa.DurationArray):
+            data = cudf.core.column.TimeDeltaColumn.from_arrow(arbitrary)
+        elif isinstance(arbitrary, pa.Date64Array):
+            raise NotImplementedError
+            data = cudf.core.column.DatetimeColumn.from_arrow(
+                arbitrary, dtype=np.dtype("M8[ms]")
+            )
+        elif isinstance(arbitrary, pa.Date32Array):
+            # No equivalent np dtype and not yet supported
+            warnings.warn(
+                "Date32 values are not yet supported so this will "
+                "be typecast to a Date64 value",
+                UserWarning,
+            )
+            data = as_column(arbitrary.cast(pa.int32())).astype("M8[ms]")
+        elif isinstance(arbitrary, pa.BooleanArray):
+            # Arrow uses 1 bit per value while we use int8
+            dtype = np.dtype(np.bool)
+            arbitrary = arbitrary.cast(pa.int8())
+            data = cudf.core.column.NumericalColumn.from_arrow(
+                arbitrary, dtype=dtype
+            )
+        elif isinstance(arbitrary, pa.ListArray):
+            data = cudf.core.column.ListColumn.from_arrow(arbitrary)
+        else:
+            data = cudf.core.column.NumericalColumn.from_arrow(arbitrary)
+
+    elif isinstance(arbitrary, pa.ChunkedArray):
+        gpu_cols = [
+            as_column(chunk, dtype=dtype) for chunk in arbitrary.chunks
+        ]
+
+        if dtype:
+            new_dtype = dtype
+        else:
+            pa_type = arbitrary.type
+            if pa.types.is_dictionary(pa_type):
+                new_dtype = "category"
+            else:
+                new_dtype = np.dtype(pa_type.to_pandas_dtype())
+
+        data = ColumnBase._concat(gpu_cols, dtype=new_dtype)
+>>>>>>> fd94d283b3c828fa49105acc13a802072e27069c
 
     elif isinstance(arbitrary, (pd.Series, pd.Categorical)):
         if isinstance(arbitrary, pd.Series) and isinstance(
@@ -1460,7 +1548,7 @@ def as_column(arbitrary, nan_as_null=None, dtype=None, length=None):
 
         if arb_dtype.kind == "M":
 
-            time_unit, _ = np.datetime_data(arbitrary.dtype)
+            time_unit = get_time_unit(arbitrary)
             cast_dtype = time_unit in ("D", "W", "M", "Y")
 
             if cast_dtype:
@@ -1476,6 +1564,26 @@ def as_column(arbitrary, nan_as_null=None, dtype=None, length=None):
                 mask = data.mask
 
             data = cudf.core.column.datetime.DatetimeColumn(
+                data=buffer, mask=mask, dtype=arbitrary.dtype
+            )
+        elif arb_dtype.kind == "m":
+
+            time_unit = get_time_unit(arbitrary)
+            cast_dtype = time_unit in ("D", "W", "M", "Y")
+
+            if cast_dtype:
+                arbitrary = arbitrary.astype(np.dtype("timedelta64[s]"))
+
+            buffer = Buffer(arbitrary.view("|u1"))
+            mask = None
+            if nan_as_null is None or nan_as_null is True:
+                data = as_column(
+                    buffer, dtype=arbitrary.dtype, nan_as_null=nan_as_null
+                )
+                data = utils.time_col_replace_nulls(data)
+                mask = data.mask
+
+            data = cudf.core.column.timedelta.TimeDeltaColumn(
                 data=buffer, mask=mask, dtype=arbitrary.dtype
             )
         elif arb_dtype.kind in ("O", "U"):